+++ packages/gasket-plugin-engine/test/execWaterfall.test.js
@@ -18,19 +18,26 @@ describe('The execWaterfall method', () => {
       }
     };
 
-    jest.resetModules();
-
     jest
-      .doMock('@gasket/a-plugin', () => pluginA, { virtual: true })
-      .doMock('@gasket/b-plugin', () => pluginB, { virtual: true });
+      .doMock('@gasket/testa-plugin', () => pluginA, { virtual: true })
+      .doMock('@gasket/testb-plugin', () => pluginB, { virtual: true });
 
     const PluginEngine = require('..');
+    const Resolver = require('../lib/resolver');
+    jest.spyOn(Resolver.prototype, 'tryResolve').mockImplementation(arg => {
+      return `${process.cwd()}/node_modules/${arg}`;
+    });
 
     engine = new PluginEngine({
       plugins: {
-        add: ['a', 'b']
+        add: ['testa', 'testb']
       }
     });
+  });
+
+  afterEach(() => {
+    jest.resetModules();
+    jest.restoreAllMocks();
   });
 
   it('sequentially transforms a value', async () => {
@@ -43,8 +50,8 @@ describe('The execWaterfall method', () => {
 
     const result = await engine.execWaterfall('eventA', 5, otherArg);
 
-    expect(pluginA.hooks.eventA).toBeCalledWith(engine, 5, otherArg);
-    expect(pluginB.hooks.eventA).toBeCalledWith(engine, 35, otherArg);
+    expect(pluginA.hooks.eventA).toHaveBeenCalledWith(engine, 5, otherArg);
+    expect(pluginB.hooks.eventA).toHaveBeenCalledWith(engine, 35, otherArg);
     expect(result).toEqual(39);
   });