+++ packages/gasket-resolve/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `@gasket/resolve`
 
+###
+
+- Including `package-identifier.js` and exposing new relative path resolve functions.
+
 ### 1.3.0
 
 - [#8] Allow presets to extend other presets